services/user-service/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
cryptography==41.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-cors==3.0.10